Enhance drcachesim documentation (#2732)
Improves the drcachesim documentation by adding example output for each
tool, adding new sections on tracing a subset of execution, creating new
analysis tools, highlighting offline support, etc.

+++ clients/drcachesim/launcher.cpp
@@ -87,7 +87,8 @@ signal_handler(int sig, siginfo_t *info, void *cxt)
     (void)res; // Work around compiler warnings.
     // Terminate child in case shell didn't already send this there.
     // It's up to the child to terminate grandchildren not already notified.
-    kill(child, SIGINT);
+    if (child != 0)
+        kill(child, SIGINT);
     // Destroy pipe file if it's open.
     if (analyzer != NULL)
         delete analyzer;
